requirements.txt
absl-py==0.9.0
adal==1.2.4
alembic==1.4.2
aniso8601==8.0.0
APScheduler==3.5.0
arrow==0.15.5
astunparse==1.6.3
atomicwrites==1.4.0
attrs==19.3.0
Authlib==0.14.1
Automat==20.2.0
brotlipy==0.7.0
cached-property==1.5.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
constantly==15.1.0
coverage==4.5.4
cryptography==2.9.2
cssselect==1.1.0
cycler==0.10.0
Cython==0.29.16
dnspython==1.16.0
docker==4.2.0
docker-compose==1.25.5
docker-py==1.10.6
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
dogpile.cache==0.6.5
gast==0.3.3
google-pasta==0.2.0
greenlet==0.4.15
grpcio==1.30.0
h5py==2.10.0
hyperlink==19.0.0
idna==2.10
importlib-metadata==1.5.0
incremental==17.5.0
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.3
joblib==0.14.1
jsonschema==3.2.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lxml==4.5.0
Mako==1.1.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
monotonic==1.5
more-itertools==8.2.0
msrest==0.6.11
msrestazure==0.6.3
nose==1.3.4
numpy==1.18.2
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.3
pandas==1.0.3
paramiko==2.7.1
parsel==1.5.2
pluggy==0.13.1
Protego==0.1.16
protobuf==3.12.2
psycopg2==2.8.4
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyJWT==1.7.1
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pypiwin32==223
pyrsistent==0.15.7
PySocks==1.7.1
pytest==5.0.1
pytest-cov==2.7.1
pytest-docker-compose==3.0.0
pytest-html==1.22.0
pytest-metadata==1.8.0
pytest-ordering==0.6
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==3.0.0
python-socketio==2.1.0
pytz==2019.3
pywin32==227
PyYAML==5.3.1
queuelib==1.5.0
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.1
service-identity==18.1.0
shub==2.10.0
six==1.15.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
texttable==1.6.2
toml==0.10.1
tqdm==4.46.0
trakt.py==3.2.0
Twisted==20.3.0
typing==3.7.4.1
tzlocal==2.0.0
uamqp==1.2.6
urllib3==1.25.9
w3lib==1.21.0
wcwidth==0.1.9
websocket-client==0.57.0
Werkzeug==0.16.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
xgboost==1.0.2
zipp==3.1.0
zope.interface==5.1.0
